Setup Husky

+++ contracts/interfaces/revest/IOutputReceiver.sol
@@ -3,27 +3,24 @@
 pragma solidity >=0.8.0;
 
 import "./IRegistryProvider.sol";
-import '@openzeppelin/contracts/utils/introspection/IERC165.sol';
-
+import "@openzeppelin/contracts/utils/introspection/IERC165.sol";
 
 /**
  * @title Provider interface for Revest FNFTs
  */
 interface IOutputReceiver is IRegistryProvider, IERC165 {
-
     function receiveRevestOutput(
-        uint fnftId,
+        uint256 fnftId,
         address asset,
         address payable owner,
-        uint quantity
+        uint256 quantity
     ) external;
 
-    function getCustomMetadata(uint fnftId) external view returns (string memory);
+    function getCustomMetadata(uint256 fnftId) external view returns (string memory);
 
-    function getValue(uint fnftId) external view returns (uint);
+    function getValue(uint256 fnftId) external view returns (uint256);
 
-    function getAsset(uint fnftId) external view returns (address);
+    function getAsset(uint256 fnftId) external view returns (address);
 
-    function getOutputDisplayValues(uint fnftId) external view returns (bytes memory);
-
+    function getOutputDisplayValues(uint256 fnftId) external view returns (bytes memory);
 }
