+++ contracts/treasury/SolidexBalanceAdapter.sol
@@ -1,23 +1,28 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.0;
 
-
 import "./TreasuryTracker.sol";
 
 interface LpDepositor {
-    function userBalances(address _staker, address _token) external view returns (uint);
-
+    function userBalances(address _staker, address _token)
+        external
+        view
+        returns (uint256);
 }
 
 contract SolidexBalanceAdapter is IBalanceAdapter {
-    
     LpDepositor immutable lpDepositor;
-    
+
     constructor(address _lpDepositor) {
         lpDepositor = LpDepositor(_lpDepositor);
     }
-    
-    function balance(address _holder, address _token) external view override returns(uint) {
+
+    function balance(address _holder, address _token)
+        external
+        view
+        override
+        returns (uint256)
+    {
         return lpDepositor.userBalances(_holder, _token);
     }
 }
